Split literal constraints detection from `CanonicalizeMfp`.
We are planning to call `CanonicalizeMfp` much more times, so we
shouldn't do the expensive literal constraints detection along with
canonicalizing an MFP.
