+++ src/transform/src/lib.rs
@@ -106,6 +106,7 @@ pub mod demand;
 pub mod fold_constants;
 pub mod fusion;
 pub mod join_implementation;
+pub mod literal_constraints;
 pub mod literal_lifting;
 pub mod monotonic;
 pub mod nonnull_requirements;
@@ -466,11 +467,11 @@ impl Optimizer {
         // Implementation transformations
         let transforms: Vec<Box<dyn crate::Transform>> = vec![
             // It's important that
-            // - there is a run of CanonicalizeMfp before JoinImplementation lifts away the Filters
-            //   from the Gets;
-            // - there is no RelationCSE between this CanonicalizeMfp and JoinImplementation,
+            // - there is a run of LiteralConstraints before JoinImplementation lifts away the
+            //   Filters from the Gets;
+            // - there is no RelationCSE between this LiteralConstraints and JoinImplementation,
             //   because that could move an IndexedFilter behind a Get.
-            Box::new(crate::canonicalize_mfp::CanonicalizeMfp),
+            Box::new(crate::literal_constraints::LiteralConstraints),
             Box::new(crate::Fixpoint {
                 limit: 100,
                 transforms: vec![
@@ -510,6 +511,7 @@ impl Optimizer {
             Box::new(crate::Fixpoint {
                 limit: 100,
                 transforms: vec![
+                    Box::new(crate::canonicalize_mfp::CanonicalizeMfp),
                     // Remove threshold operators which have no effect.
                     Box::new(crate::threshold_elision::ThresholdElision),
                     // Projection pushdown may unblock fusing joins and unions.
